requirements.txt
aioredis==1.3.1
arrow==1.2.3
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.1
Automat==22.10.0
beautifulsoup4==4.11.1
binaryornot==0.4.4
certifi==2022.9.14
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
chardet==5.0.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
constantly==15.1.0
cookiecutter==2.1.1
cryptography==39.0.0
cssbeautifier==1.14.6
daphne==4.0.0
Django==4.1.1
django-browser-reload==1.6.0
django-crispy-forms==1.14.0
django-phonenumber-field==7.0.0
djlint==1.16.0
EditorConfig==0.12.3
fast-enum==1.3.0
flake8==5.0.4
hiredis==2.1.1
html-tag-names==0.1.2
html-void-elements==0.1.0
hyperlink==21.0.0
idna==3.4
importlib-metadata==4.12.0
incremental==22.10.0
Jinja2==3.1.2
jinja2-time==0.2.0
jsbeautifier==1.14.6
Logbook==1.5.3
MarkupSafe==2.1.1
mccabe==0.7.0
msgpack==1.0.4
orjson==3.8.0
pathspec==0.10.1
phonenumbers==8.12.56
Pillow==9.2.0
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
pyOpenSSL==23.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.6
python-slugify==6.1.2
PyYAML==6.0
redis==4.4.2
regex==2022.9.13
requests==2.28.1
service-identity==21.1.0
single-version==1.5.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
text-unidecode==1.3
tomli==2.0.1
tornado==6.2
tqdm==4.64.1
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.4.0
tzdata==2022.2
Unidecode==1.3.6
urllib3==1.26.12
vietnam-provinces==0.4.3
yapf==0.32.0
zipp==3.8.1
zope.interface==5.5.2
